useWorkplaceManagement.ts로 근무지 관리 로직 hook 분리
